# artifacts/api-server/src/routes/opportunities.ts
import { Router } from "express";
import type { Request, Response } from "express";
import { requireAuth } from "../lib/session-middleware";
import { query, queryOne } from "../lib/db";

const router = Router();
router.use(requireAuth);

router.get("/", async (req: Request, res: Response) => {
  const { stage, tier, human_action_required, limit } = req.query;
  const where: string[] = ["stage <> 'dismissed'", "status <> 'archived'"];
  const params: unknown[] = [];

  if (stage) { params.push(stage); where.push(`stage = $${params.length}`); }
  if (tier) { params.push(tier); where.push(`tier = $${params.length}`); }
  if (human_action_required === "true") where.push("human_action_required = true");

  const lim = Math.min(Number(limit ?? 500), 1000);
  const rows = await query(
    `select * from opportunities where ${where.join(" and ")} order by (deadline is null), deadline asc limit ${lim}`,
    params
  );
  res.json(rows);
});

router.get("/:id", async (req: Request, res: Response) => {
  const row = await queryOne(`select * from opportunities where id = $1`, [req.params.id]);
  if (!row) { res.status(404).json({ error: "Not found" }); return; }
  res.json(row);
});

router.post("/:id/action", async (req: Request, res: Response) => {
  const { action, stage } = req.body ?? {};
  const { id } = req.params;

  if (action === "pursue") {
    await query(
      `update opportunities set tier='pursue', stage='analysis', human_action_required=false, updated_at=now() where id=$1`,
      [id]
    );
  } else if (action === "dismiss") {
    await query(
      `update opportunities set tier='dismiss', stage='dismissed', human_action_required=false, updated_at=now() where id=$1`,
      [id]
    );
  } else if (action === "advance" && stage) {
    await query(
      `update opportunities set stage=$1, human_action_required=false, updated_at=now() where id=$2`,
      [stage, id]
    );
  } else {
    res.status(400).json({ error: "Invalid action" });
    return;
  }
  res.json({ ok: true });
});

export default router;